+++ database/mapSolarSystems.sql
@@ -1,6 +1,6 @@
 -- MySQL dump 10.13  Distrib 5.6.12, for Linux (x86_64)
 --
--- Host: localhost    Database: sdegalatea1
+-- Host: localhost    Database: sdevanguard1
 -- ------------------------------------------------------
 -- Server version	5.6.12
 
@@ -78,4 +78,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2015-08-25  1:48:15
+-- Dump completed on 2015-09-29 17:45:33